--- src/main/java/com/inq/kbus/KbusApplication.java
@@ -1,13 +1,0 @@
-package com.inq.kbus;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KbusApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(KbusApplication.class, args);
-	}
-
-}
