--- src/main/java/com/inq/kbus/domain/booking/Booking.java
@@ -1,34 +1,0 @@
-package com.inq.kbus.domain.booking;
-
-import com.inq.kbus.domain.member.Member;
-import com.inq.kbus.domain.place.Place;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-@Entity @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Booking {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "place_id")
-    private Place place;
-
-    private LocalDate date;
-
-    private LocalTime startTime;
-
-    private LocalTime endTime;
-}
