--- src/main/java/com/inq/kbus/domain/department/repository/DepartmentRepository.java
@@ -1,8 +1,0 @@
-package com.inq.kbus.domain.department.repository;
-
-import com.inq.kbus.domain.department.Department;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DepartmentRepository extends JpaRepository<Department, Long> {
-
-}
