--- src/main/java/com/inq/kbus/domain/member/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.inq.kbus.domain.member.repository;
-
-import com.inq.kbus.domain.member.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-
-}
