--- src/main/java/com/inq/kbus/domain/place/Area.java
@@ -1,5 +1,0 @@
-package com.inq.kbus.domain.place;
-
-public enum Area {
-    ONE, TWO, THREE
-}
