--- src/main/java/com/inq/kbus/domain/place/Place.java
@@ -1,26 +1,0 @@
-package com.inq.kbus.domain.place;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Place {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private Sector sector;
-
-    @Enumerated(EnumType.STRING)
-    private Area area;
-
-    private String pName;
-
-    private Integer floor;
-}
