--- src/main/java/com/inq/kbus/domain/place/Sector.java
@@ -1,5 +1,0 @@
-package com.inq.kbus.domain.place;
-
-public enum Sector {
-    ONE, TWO, THREE, FOUR
-}
