--- src/main/java/com/inq/kbus/web/member/dto/response/SimpleMemberCreatedResponse.java
@@ -1,13 +1,0 @@
-package com.inq.kbus.web.member.dto.response;
-
-import lombok.Data;
-
-@Data
-public class SimpleMemberCreatedResponse {
-
-    private Long id;
-
-    public SimpleMemberCreatedResponse(Long id) {
-        this.id = id;
-    }
-}
