+++ src/main/java/com/inq/kubs/KbusApplication.java
@@ -1,0 +1,13 @@
+package com.inq.kubs;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KbusApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(KbusApplication.class, args);
+	}
+
+}
