+++ src/main/java/com/inq/kubs/domain/department/repository/DepartmentRepository.java
@@ -1,0 +1,8 @@
+package com.inq.kubs.domain.department.repository;
+
+import com.inq.kubs.domain.department.Department;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
+
+}
