+++ src/main/java/com/inq/kubs/domain/email/MailDto.java
@@ -1,0 +1,17 @@
+package com.inq.kubs.domain.email;
+
+import lombok.Data;
+
+@Data
+public class MailDto {
+
+    private String address;
+    private String title;
+    private String content;
+
+    public MailDto(String address, String title, String content) {
+        this.address = address;
+        this.title = title;
+        this.content = content;
+    }
+}
