+++ src/main/java/com/inq/kubs/domain/member/repository/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.inq.kubs.domain.member.repository;
+
+import com.inq.kubs.domain.member.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+
+}
