+++ src/main/java/com/inq/kubs/domain/place/Area.java
@@ -1,0 +1,5 @@
+package com.inq.kubs.domain.place;
+
+public enum Area {
+    ONE, TWO, THREE
+}
