+++ src/main/java/com/inq/kubs/domain/place/Place.java
@@ -1,0 +1,26 @@
+package com.inq.kubs.domain.place;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Place {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Enumerated(EnumType.STRING)
+    private Sector sector;
+
+    @Enumerated(EnumType.STRING)
+    private Area area;
+
+    private String pName;
+
+    private Integer floor;
+}
