+++ src/main/java/com/inq/kubs/domain/place/Sector.java
@@ -1,0 +1,5 @@
+package com.inq.kubs.domain.place;
+
+public enum Sector {
+    ONE, TWO, THREE, FOUR
+}
