+++ src/main/java/com/inq/kubs/web/exception/ApiExceptionHandler.java
@@ -1,0 +1,21 @@
+package com.inq.kubs.web.exception;
+
+import com.inq.kubs.web.member.MemberController;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice(assignableTypes = {
+        MemberController.class
+})
+public class ApiExceptionHandler {
+
+    @ExceptionHandler(KubsException.class)
+    public ResponseEntity<ErrorResponse> handleException(KubsException e) {
+
+        ErrorType errorType = e.getErrorType();
+        ErrorResponse errorResponse = new ErrorResponse(false, errorType.getCode(), errorType.getMessage());
+
+        return new ResponseEntity<>(errorResponse, errorType.getHttpStatus());
+    }
+}
