+++ src/main/java/com/inq/kubs/web/exception/ErrorResponse.java
@@ -1,0 +1,17 @@
+package com.inq.kubs.web.exception;
+
+import lombok.Data;
+
+@Data
+public class ErrorResponse {
+
+    private boolean success;
+    private String code;
+    private String message;
+
+    public ErrorResponse(boolean success, String code, String message) {
+        this.success = success;
+        this.code = code;
+        this.message = message;
+    }
+}
