+++ src/main/java/com/inq/kubs/web/exception/ErrorType.java
@@ -1,0 +1,20 @@
+package com.inq.kubs.web.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public enum ErrorType {
+
+    NOT_EXIST_PK("Not Found", "존재하지 않는 PK 입니다.", HttpStatus.NOT_FOUND);
+
+    private final String code;
+    private final String message;
+    private final HttpStatus httpStatus;
+
+    ErrorType(String code, String message, HttpStatus httpStatus) {
+        this.code = code;
+        this.message = message;
+        this.httpStatus = httpStatus;
+    }
+}
