+++ src/main/java/com/inq/kubs/web/member/dto/response/SimpleMemberCreatedResponse.java
@@ -1,0 +1,13 @@
+package com.inq.kubs.web.member.dto.response;
+
+import lombok.Data;
+
+@Data
+public class SimpleMemberCreatedResponse {
+
+    private Long id;
+
+    public SimpleMemberCreatedResponse(Long id) {
+        this.id = id;
+    }
+}
