--- src/test/java/com/inq/kbus/KbusApplicationTests.java
@@ -1,13 +1,0 @@
-package com.inq.kbus;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class KbusApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
