+++ src/test/java/com/inq/kubs/KbusApplicationTests.java
@@ -1,0 +1,13 @@
+package com.inq.kubs;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class KbusApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
